+++ _pytest/config.py
@@ -996,6 +996,7 @@ class Config(object):
                                  "(are you using python -O?)\n")
 
     def _preparse(self, args, addopts=True):
+        import pytest
         self._initini(args)
         if addopts:
             args[:] = shlex.split(os.environ.get('PYTEST_ADDOPTS', '')) + args
@@ -1007,7 +1008,10 @@ class Config(object):
         self.pluginmanager.load_setuptools_entrypoints(entrypoint_name)
         self.pluginmanager.consider_env()
         self.known_args_namespace = ns = self._parser.parse_known_args(args, namespace=self.option.copy())
-        if self.known_args_namespace.confcutdir is None and self.inifile:
+        confcutdir = self.known_args_namespace.confcutdir
+        if confcutdir and not os.path.isdir(confcutdir):
+            raise pytest.UsageError('--confcutdir must be a directory, given: {0}'.format(confcutdir))
+        if confcutdir is None and self.inifile:
             confcutdir = py.path.local(self.inifile).dirname
             self.known_args_namespace.confcutdir = confcutdir
         try: